Fix directory deletion
